Add try/except to support more ceph client versions
The share creation process attempts to create a volume with a "mode"
which may not be supported by the ceph client version installed on a
system. When running rocky or master with ceph minic and python client
version "13.2.4" the manila-share service will stacktrace with the
following output:

> https://pasted.tech/pastes/d678e0c00f8be82d60bd113359b516f4551ed3a6.raw ]

This change simply adds a try/except with a TypeError falling back to a
volume create call without the "mode" keyword argument allowing the
share create to proceed normally.

Change-Id: I9a88d3a11f099386b8a390e6aea3688280d9a2af
Signed-off-by: cloudnull <kevin@cloudnull.com>